requirements.txt
pandas>=2.0.0
scikit-learn>=1.0.0
pathlib>=1.0.1
typing>=3.7.4
numpy>=1.24.0
openai>=1.12.0
anthropic>=0.18.1
python-dotenv>=1.0.0
requests>=2.31.0 